+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 ## [0.1.1] - 2026-01-15
 
 ### Fixed
-- Fix `Application.app_dir/2` error during initial compilation of projects using c3nif
+- Fix `Application.app_dir/2` errors during initial compilation of projects using c3nif
 
 ## [Unreleased]